Moved size back to a Long instead of long.

+++ src/test/java/com/flightstats/hub/spoke/ContentMarshallerTest.java
@@ -66,7 +66,7 @@ public class ContentMarshallerTest {
         Content cycled = ContentMarshaller.toContent(ContentMarshaller.toBytes(content), content.getContentKey().get());
         assertTrue(content.equals(cycled));
         assertEquals(size, cycled.getData().length);
-        assertEquals(size, cycled.getSize());
+        assertEquals(size, cycled.getSize().longValue());
     }
 
     private static Content getContent(int size) {
